Python_个人项目/人工智能/训练一个猜测你喜好类型图片的AI/requirements.txt
torch
torchvision
pillow
numpy
psutil
markupsafe==2.0.1
flask